remove threat import cycles
