+++ ddtrace/appsec/_asm_request_context.py
@@ -662,12 +662,6 @@ def _set_headers(span: Span, headers: Any, kind: str, only_asm_enabled: bool = F
 
 
 def asm_listen():
-    from ddtrace.appsec._handlers import listen
-    from ddtrace.appsec._trace_utils import listen as trace_listen
-
-    listen()
-    trace_listen()
-
     core.on("flask.finalize_request.post", _set_headers_and_response)
     core.on("flask.wrapped_view", _on_wrapped_view, "callbacks")
     core.on("flask._patched_request", _on_pre_tracedrequest)
