+++ ddtrace/appsec/_processor.py
@@ -86,7 +86,7 @@ class AppSecSpanProcessor(SpanProcessor):
         return self._ddwaf is not None
 
     def __post_init__(self) -> None:
-        from ddtrace.appsec import load_appsec
+        from ddtrace.appsec._listeners import load_appsec
 
         load_appsec()
         self.obfuscation_parameter_key_regexp = asm_config._asm_obfuscation_parameter_key_regexp.encode()
